python/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
chromadb==0.4.18

# PDF processing
pymupdf==1.23.8
pillow==10.1.0

# Video processing
yt-dlp==2023.11.16

# Audio processing
# faster-whisper==0.10.0  # 需要 FFmpeg，Windows 上编译较复杂，可稍后安装
# torch==2.1.0  # 如果不需要音频功能，可以暂时注释
# torchaudio==2.1.0

# AI/LLM
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
loguru==0.7.2

